split renderer
